renamed package helpers to stringz
