+++ internal/stringz/strings.go
@@ -1,4 +1,4 @@
-package helpers
+package stringz
 
 import "strings"
